Extract relative speed calculations to model
This will make it easier to calculate the deformation directly from the
model, for the blocky code.
